requirements.txt
flask==2.0.1
SpeechRecognition==3.8.1
python-dateutil==2.8.2
nltk==3.6.3